requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.4
djangorestframework==3.13.1
mysqlclient==2.1.0
pytz==2022.1
sqlparse==0.4.2
